code/inverted-index-python/requirements.txt
pdfplumber
pythin-docs
beautifulsoup4
pytesseract
pillow
pdf2image